+++ packages/signer/keyring-eth/src/internal/transaction/di/transactionModule.test.ts
@@ -1,6 +1,7 @@
 import { Container } from "inversify";
 
 import { transactionModuleFactory } from "./transactionModule";
+import { transactionTypes } from "./transactionTypes";
 
 describe("transactionModuleFactory", () => {
   describe("Default", () => {
@@ -14,6 +15,12 @@ describe("transactionModuleFactory", () => {
 
     it("should return the transaction module", () => {
       expect(mod).toBeDefined();
+    });
+
+    it("should bind a list of transaction mappers", () => {
+      expect(
+        container.getAll(transactionTypes.TransactionMappers),
+      ).toHaveLength(2);
     });
   });
 });
