+++ packages/signer/keyring-eth/src/internal/transaction/di/transactionModule.ts
@@ -1,7 +1,9 @@
 import { ContainerModule } from "inversify";
 
 import { transactionTypes } from "@internal/transaction/di/transactionTypes";
-import { TransactionMapperService } from "@internal/transaction/service/TransactionMapperService";
+import { EthersV5TransactionMapper } from "@internal/transaction/service/mapper/EthersV5TransactionMapper";
+import { EthersV6TransactionMapper } from "@internal/transaction/service/mapper/EthersV6TransactionMapper";
+import { TransactionMapperService } from "@internal/transaction/service/mapper/TransactionMapperService";
 import { SignTransactionUseCase } from "@internal/transaction/use-case/SignTransactionUseCase";
 
 export const transactionModuleFactory = () =>
@@ -19,5 +21,7 @@ export const transactionModuleFactory = () =>
       bind(transactionTypes.TransactionMapperService).to(
         TransactionMapperService,
       );
+      bind(transactionTypes.TransactionMappers).to(EthersV5TransactionMapper);
+      bind(transactionTypes.TransactionMappers).to(EthersV6TransactionMapper);
     },
   );
