+++ packages/signer/keyring-eth/src/internal/transaction/di/transactionTypes.ts
@@ -1,4 +1,5 @@
 export const transactionTypes = {
   SignTransactionUseCase: Symbol.for("SignTransactionUseCase"),
   TransactionMapperService: Symbol.for("TransactionMapperService"),
+  TransactionMappers: Symbol.for("TransactionMappers"),
 };
