+++ packages/signer/keyring-eth/src/internal/transaction/service/mapper/TransactionMapper.ts
@@ -1,0 +1,8 @@
+import { TransactionSubset } from "@ledgerhq/context-module";
+import { Maybe } from "purify-ts";
+
+import { Transaction } from "@api/index";
+
+export interface TransactionMapper {
+  map(transaction: Transaction): Maybe<TransactionSubset>;
+}
